+++ shared/random_number.c
@@ -1,10 +1,10 @@
 #include "random_number.h"
 
-int random_number(int min, int max) {
+short random_number(short min, short max) {
     if((max-min) == 0){
         return min;
     }else{
-        int num = rand() % (max-min) + min;
+        short num = rand() % (max-min) + min;
         return num;
     }
 }
